Improving location form formatting
