+++ client/src/pages/rooms/index.js
@@ -18,7 +18,7 @@ import { useMediaQuery } from 'react-responsive';
 import API from "../../utils/api";
 import "./style.css";
 import LocationsList from "../../components/LocationsList";
-import LocationForm from "../../components/LocationFormUpdate";
+import LocationForm from "../../components/LocationForm";
 import FeatureForm from "../../components/FeatureForm";
 import RoomForm from "../../components/RoomForm";
